Restructure package for modularity
